Отработка замечаний ревью: - ❌ соблюдены расположение всех элементов, отступы, размеры кнопок, размеры рамки и скруглений - ❌ изображения не деформируются, не обрезаются - ❌ После выбора ответа на вопрос через 1 секунду автоматически появляется следующий вопрос и рамка ответа исчезает И рекомендаций: - 🍏В макете статус бар белый, а у тебя черный. - 🍏Обычно используем такой порядок: сначала перечисляем свойства класса, затем функции. При этом также делим на приватные и публичные блоки. Функции перечисляют в таком порядке: initы, overrideы, блоки публичных и приватных функций, и далее блок c аннотацией @IBAction - 🍏 Не забывай следить за отступами, общепринятые отступы помогают читать код быстрее. В общем случае поможет выделить весь файл cmd+A и зажать control+i - 🍪Пользователь может случайно прожать два раза кнопку ответа или даже много и быстро понажимать на кнопки. Это может привести к неверному увеличению счетчика, или лишнего вызова алерта с результатом.
