+++ MovieQuiz/Presentation/Base.lproj/Main.storyboard
@@ -45,7 +45,7 @@
                                             </label>
                                         </subviews>
                                     </stackView>
-                                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="ux8-TB-IEV" userLabel="Preview Image">
+                                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="ux8-TB-IEV" userLabel="Preview Image">
                                         <rect key="frame" x="0.0" y="43.5" width="374" height="561"/>
                                         <color key="backgroundColor" name="YP White"/>
                                         <constraints>
@@ -115,7 +115,9 @@
                     <connections>
                         <outlet property="counterLabel" destination="SGX-M7-aA8" id="ll6-TM-ncS"/>
                         <outlet property="imageView" destination="ux8-TB-IEV" id="ooY-kk-LJD"/>
+                        <outlet property="noButton" destination="TiO-bS-0IR" id="c3u-p4-6mc"/>
                         <outlet property="textLabel" destination="Mei-uk-yi1" id="6rH-Jb-wqN"/>
+                        <outlet property="yesButton" destination="h16-Mx-PQj" id="kSq-kK-DbZ"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
