+++ MovieQuiz/Presentation/Models.swift
@@ -1,0 +1,35 @@
+//
+//  Models.swift
+//  MovieQuiz
+//
+
+import UIKit
+
+struct QuizQuestion {
+    // строка с названием фильма,
+    // совпадает с названием картинки афиши фильма в Assets
+    let image: String
+    // строка с вопросом о рейтинге фильма
+    let text: String = "Рейтинг этого фильма больше чем 6?"
+    // булевое значение (true, false), правильный ответ на вопрос
+    let corrcetAnswer: Bool
+}
+
+struct QuizStepViewModel {
+    // картинка с афишей фильма с типом UIImage
+    let image: UIImage
+    // вопрос о рейтинге квиза
+    let question: String
+    // строка с порядковым номером этого вопроса (ex. "1/10")
+    let questionNumber: String
+}
+
+// для состояния "Результат квиза"
+struct QuizResultsViewModel {
+    // строка с заголовком алерта
+    let title: String
+    // строка с текстом о количестве набранных очков
+    let text: String
+    // текст для кнопки алерта
+    let buttonText: String
+}
