+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -2,6 +2,11 @@ import UIKit
 
 final class MovieQuizViewController: UIViewController {
     // MARK: - Lifecycle
+    @IBOutlet private weak var imageView: UIImageView!
+    @IBOutlet private weak var textLabel: UILabel!
+    @IBOutlet private weak var counterLabel: UILabel!
+    @IBOutlet private weak var noButton: UIButton!
+    @IBOutlet private weak var yesButton: UIButton!
     // переменная с индексом текущего вопроса, начальное значение 0
     private var correctAnswers = 0
     // (по этому индексу будем искать вопрос в массиве, где индекс первого элемента 0, а не 1)
@@ -9,42 +14,33 @@ final class MovieQuizViewController: UIViewController {
     private var currentQuestionIndex = 0
     //массив вопросов
     private let questions: [QuizQuestion] = [
-        QuizQuestion(image: "The Godfather"
-                     , text: "Рейтинг этого фильма больше чем 6?"
-                     , corrcetAnswer: true),
+        QuizQuestion(image: "The Godfather",
+                     corrcetAnswer: true),
         QuizQuestion(image: "The Dark Knight",
-                     text: "Рейтинг этого фильма больше чем 6?",
                      corrcetAnswer: true),
         QuizQuestion(image: "Kill Bill",
-                     text: "Рейтинг этого фильма больше чем 6?",
                      corrcetAnswer: true),
         QuizQuestion(image: "The Avengers",
-                     text: "Рейтинг этого фильма больше чем 6?",
                      corrcetAnswer: true),
         QuizQuestion(image: "Deadpool",
-                     text: "Рейтинг этого фильма больше чем 6?",
                      corrcetAnswer: true),
         QuizQuestion(image: "The Green Knight",
-                     text: "Рейтинг этого фильма больше чем 6?",
                      corrcetAnswer: true),
         QuizQuestion(image: "Old",
-                     text: "Рейтинг этого фильма больше чем 6?",
                      corrcetAnswer: false),
         QuizQuestion(image: "The Ice Age Adventures of Buck Wild",
-                     text: "Рейтинг этого фильма больше чем 6?",
                      corrcetAnswer: false),
         QuizQuestion(image: "Tesla",
-                     text: "Рейтинг этого фильма больше чем 6?",
                      corrcetAnswer: false),
         QuizQuestion(image: "Vivarium",
-                     text: "Рейтинг этого фильма больше чем 6?",
                      corrcetAnswer: false)
     ]
-    
     // константа с кнопкой для системного алерта
-  
     override func viewDidLoad() {
         show(quiz: convert(model: questions[currentQuestionIndex]))
+        imageView.layer.masksToBounds = true // даём разрешение на рисование рамки
+        imageView.layer.borderWidth = 8 // толщина рамки
+        imageView.layer.cornerRadius = 20 // радиус скругления картинки
         super.viewDidLoad()
     }
     // метод вызывается, когда пользователь нажимает на кнопку "Нет"
@@ -59,10 +55,6 @@ final class MovieQuizViewController: UIViewController {
         let givenAnswer = true
         showAnswerResult(isCorrect: givenAnswer == currentQuestion.corrcetAnswer)
     }
-    @IBOutlet private weak var imageView: UIImageView!
-    @IBOutlet private weak var textLabel: UILabel!
-    @IBOutlet private weak var counterLabel: UILabel!
-    
     // приватный метод конвертации, который принимает моковый вопрос и возвращает вью модель для главного экрана
     private func convert (model: QuizQuestion) -> QuizStepViewModel {
         return QuizStepViewModel(
@@ -70,14 +62,12 @@ final class MovieQuizViewController: UIViewController {
             question: model.text,
             questionNumber:"\(currentQuestionIndex+1)/\(questions.count)")
     }
-    
     // приватный метод вывода на экран вопроса, который принимает на вход вью модель вопроса и ничего не возвращает
     private func show(quiz step: QuizStepViewModel) {
         imageView.image = step.image
         textLabel.text = step.question
         counterLabel.text = step.questionNumber
     }
-    
     // приватный метод для показа результатов раунда квиза
     // принимает вью модель QuizResultsViewModel и ничего не возвращает
     private func show(quiz result: QuizResultsViewModel) {
@@ -101,14 +91,16 @@ final class MovieQuizViewController: UIViewController {
     
     private func showAnswerResult(isCorrect: Bool) {
         if isCorrect { // 1
-               correctAnswers += 1 // 2
-           }
-        imageView.layer.masksToBounds = true // даём разрешение на рисование рамки
-        imageView.layer.borderWidth = 1 // толщина рамки
+            correctAnswers += 1 // 2
+        }
         imageView.layer.borderColor = isCorrect ? UIColor.ypGreen.cgColor : UIColor.ypRed.cgColor//
-        imageView.layer.cornerRadius = 6 // радиус скругления углов рамки
+        noButton.isEnabled = false
+        yesButton.isEnabled = false
         DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
             self.showNextQuestionOrResults()
+            self.imageView.layer.borderColor = UIColor.clear.cgColor
+            self.noButton.isEnabled = true
+            self.yesButton.isEnabled = true
         }
     }
     // приватный метод, который содержит логику перехода в один из сценариев
@@ -117,12 +109,13 @@ final class MovieQuizViewController: UIViewController {
         if currentQuestionIndex == questions.count - 1 {
             // идём в состояние "Результат квиза"
             let text = "Ваш результат: \(correctAnswers)/10"
-                   let viewModel = QuizResultsViewModel(
-                       title: "Этот раунд окончен!",
-                       text: text,
-                       buttonText: "Сыграть ещё раз")
-                   show(quiz: viewModel)
-        } else { 
+            let viewModel = QuizResultsViewModel(
+                title: "Этот раунд окончен!",
+                text: text,
+                buttonText: "Сыграть ещё раз")
+            show(quiz: viewModel)
+            correctAnswers = 0
+        } else {
             currentQuestionIndex += 1
             let nextQuestion = questions[currentQuestionIndex]
             let viewModel = convert(model: nextQuestion)
@@ -131,41 +124,3 @@ final class MovieQuizViewController: UIViewController {
         }
     }
 }
-
-struct QuizQuestion {
-    // строка с названием фильма,
-    // совпадает с названием картинки афиши фильма в Assets
-    let image: String
-    // строка с вопросом о рейтинге фильма
-    let text: String
-    // булевое значение (true, false), правильный ответ на вопрос
-    let corrcetAnswer: Bool
-}
-
-struct QuizStepViewModel {
-  // картинка с афишей фильма с типом UIImage
-  let image: UIImage
-  // вопрос о рейтинге квиза
-  let question: String
-  // строка с порядковым номером этого вопроса (ex. "1/10")
-  let questionNumber: String
-}
-
-// для состояния "Результат квиза"
-struct QuizResultsViewModel {
-  // строка с заголовком алерта
-  let title: String
-  // строка с текстом о количестве набранных очков
-  let text: String
-  // текст для кнопки алерта
-  let buttonText: String
-}
-
-/*
- Mock-данные
- 
- Картинка: Vivarium
- Настоящий рейтинг: 5,8
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: НЕТ
- */
