+++ MovieQuiz/Resources/Info.plist
@@ -26,5 +26,7 @@
 			</array>
 		</dict>
 	</dict>
+	<key>UIViewControllerBasedStatusBarAppearance</key>
+	<false/>
 </dict>
 </plist>
